% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{correlacion}
\alias{correlacion}
\title{Funcion que calcula la matriz de correlacion de un dataframe entre todas las variables numéricas.}
\usage{
correlacion(df)
}
\arguments{
\item{df:}{dataframe}
}
\value{
: lista con el df de correlacion entre las variables numericas y la matriz de correlacion
}
\description{
Funcion que calcula la matriz de correlacion de un dataframe entre todas las variables numéricas.
}
