% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{covarianza}
\alias{covarianza}
\title{Funcion para calcular la covarianza de dos vectores}
\usage{
covarianza(x, y)
}
\arguments{
\item{x:}{primer vector de datos}

\item{y:}{segundo vector de datos}
}
\value{
: covarianza de los vectores
}
\description{
Funcion para calcular la covarianza de dos vectores
}
