% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{desviacion}
\alias{desviacion}
\title{Funcion para calcular la desviacion estandar de un vector}
\usage{
desviacion(x)
}
\arguments{
\item{x:}{vector de datos}
}
\value{
: desviacion estandar del vector
}
\description{
Funcion para calcular la desviacion estandar de un vector
}
