% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizar}
\alias{discretizar}
\title{Algoritmos de discretización para un solo atributo y para un dataset completo.}
\usage{
discretizar(
  datos,
  metodo = "anchura",
  n_intervalos = NULL,
  frecuencia = NULL,
  columnas = NULL,
  verbose = 1
)
}
\arguments{
\item{datos}{Serie numérica o DataFrame sobre el que realizar la discretización.}

\item{metodo}{\code{"anchura"} (igual anchura) o \code{"frecuencia"} (igual frecuencia).}

\item{n_intervalos}{Número de intervalos a crear. Si es NULL, se calcula automáticamente como el 50\% del número de casos.}

\item{frecuencia}{Vector de frecuencias objetivo por intervalo (método "frecuencia").}

\item{columnas}{Lista de nombres de columnas a discretizar, o NULL para usar todas las columnas numéricas del dataset.}

\item{verbose}{Nivel de información que se muestra durante la ejecución.}
}
\value{
La variable discretizada (un vector/factor) si \code{datos} es una serie, o un DataFrame con las columnas discretizadas si \code{datos} es un dataset.
}
\description{
La discretización transforma una variable continua/númerica en categorías o
intervalos. Se contemplan dos métodos:
}
\details{
\itemize{
\item "anchura": igual anchura e igual número de casos por intervalo.
\item "frecuencia": igual frecuencia (número de casos) e igual anchura variable.
}
}
\section{Uso}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{discretizar(series) -> una sola columna
discretizar(df, columnas = c(...)) -> varias columnas a la vez
discretizar(df) -> todas las columnas numéricas del DataFrame
}\if{html}{\out{</div>}}
}

