% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizar_variable}
\alias{discretizar_variable}
\title{Decide qué método de discretización utilizar según el parámetro "metodo".}
\usage{
discretizar_variable(
  columna_variable,
  metodo,
  n_intervalos,
  frecuencia,
  verbose = 1
)
}
\arguments{
\item{columna_variable}{Variable numérica a discretizar.}

\item{metodo}{Método a aplicar: "anchura" o "frecuencia".}

\item{n_intervalos}{Número de intervalos deseados.}

\item{frecuencia}{Número de casos por intervalo.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
El resultado de la discretización según el método indicado.
}
\description{
Decide qué método de discretización utilizar según el parámetro "metodo".
}
