% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizarEF}
\alias{discretizarEF}
\title{Funcion para discretizar un vector o dataframe con el metodo de equal frequency}
\usage{
discretizarEF(x, num_bins, labels = FALSE)
}
\arguments{
\item{x:}{dataframe o vector}

\item{num_bins:}{número de intervalos que queremos obtener}

\item{labels:}{Booleano para indicar si se quiere que los intervalos se muestren en el array (True) o solo el indice del intervalo (False)}
}
\value{
: dataframe o vector discretizado
}
\description{
Funcion para discretizar un vector o dataframe con el metodo de equal frequency
}
