% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlacion_info.r
\name{.calc_corr_catg_num}
\alias{.calc_corr_catg_num}
\title{Welch ANOVA de un factor entre un atributo categórico y uno numérico.}
\usage{
.calc_corr_catg_num(atributo1, atributo2, verbose = 1)
}
\arguments{
\item{atributo1}{Nombre o columna del primer atributo.}

\item{atributo2}{Nombre o columna del segundo atributo.}

\item{verbose}{Nivel de información que se muestra durante la ejecución.}
}
\value{
Una lista con el estadístico F y toda la información por grupo.
}
\description{
Comprueba si existen diferencias entre las medias de una variable continua
para los distintos grupos de una variable discreta.
}
