% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlacion_info.r
\name{.calc_corr_num}
\alias{.calc_corr_num}
\title{Correlación de Pearson entre dos atributos numéricos.}
\usage{
.calc_corr_num(atributo1, atributo2, verbose = 1)
}
\arguments{
\item{atributo1}{Nombre o columna del primer atributo numérico.}

\item{atributo2}{Nombre o columna del segundo atributo numérico.}

\item{verbose}{Nivel de información que se muestra durante la ejecución.}
}
\value{
Un list con el coeficiente de correlación \code{r}, el coeficiente de determinación \code{r2}, el número de observaciones \code{n}, y los vectores \code{x} e \code{y} necesarios para pintar el gráfico de dispersión + recta de regresión.
}
\description{
Mide la intensidad y dirección de la relación LINEAL entre dos variables
numéricas. El resultado está entre -1 y 1:
}
\details{
\itemize{
\item +1 -> relación lineal positiva perfecta
\item 0 -> ausencia de relación lineal
\item -1 -> relación lineal negativa perfecta
}
}
