% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_de_variables.r
\name{.calcular_entropia}
\alias{.calcular_entropia}
\title{Calcula la entropía de una variable discreta respecto a las clases.
La entropía mide cuánta incertidumbre o diversidad hay respecto a las clases de una variable objetivo. Si todos los
ejemplos pertenecen a la misma clase, la entropía es 0 porque no existe incertidumbre. Cuanto más repartidos estén los
ejemplos entre las distintas clases, mayor será la entropía.}
\usage{
.calcular_entropia(atributo_clm, verbose)
}
\arguments{
\item{atributo_clm}{Vector de valores discretos de la variable objetivo.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
El valor de la entropía calculada.
}
\description{
Calcula la entropía de una variable discreta respecto a las clases.
La entropía mide cuánta incertidumbre o diversidad hay respecto a las clases de una variable objetivo. Si todos los
ejemplos pertenecen a la misma clase, la entropía es 0 porque no existe incertidumbre. Cuanto más repartidos estén los
ejemplos entre las distintas clases, mayor será la entropía.
}
