% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_de_variables.r
\name{.calcular_metricas_continua}
\alias{.calcular_metricas_continua}
\title{Calcula las métricas de una variable continua.}
\usage{
.calcular_metricas_continua(columna, clases, verbose = 1)
}
\arguments{
\item{columna}{Variable continua a analizar.}

\item{clases}{Columna binaria opcional con la que evaluar el AUC de las variables continuas.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
Lista con \code{varianza}. Si se proporcionan clases, también incluye \code{auc}.
}
\description{
Calcula las métricas de una variable continua.
}
