% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_de_variables.r
\name{.calcular_metricas_dataset}
\alias{.calcular_metricas_dataset}
\title{Calcula las métricas de un dataset entero.
Se calculan las métricas de cada columna indicada (o de todas las columnas).}
\usage{
.calcular_metricas_dataset(dataset, clases, atributos_clm, verbose = 1)
}
\arguments{
\item{dataset}{Datos de entrada. Un data.frame.}

\item{clases}{Columna binaria opcional con la que evaluar el AUC de las variables continuas.}

\item{atributos_clm}{Columnas específicas a analizar. Si es NULL, se analizan todas las columnas numéricas.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
El dataset con cada columna anotada con su métrica (entropía o varianza/auc).
}
\description{
Calcula las métricas de un dataset entero.
Se calculan las métricas de cada columna indicada (o de todas las columnas).
}
