% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_de_variables.r
\name{.calcular_metricas_discreta}
\alias{.calcular_metricas_discreta}
\title{Calcula las métricas de una variable discreta.}
\usage{
.calcular_metricas_discreta(columna, verbose = 1)
}
\arguments{
\item{columna}{Variable discreta a analizar.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
Lista con la clave \code{entropia} y su valor.
}
\description{
Calcula las métricas de una variable discreta.
}
