% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_de_variables.r
\name{.calcular_metricas_variable}
\alias{.calcular_metricas_variable}
\title{Decide qué métrica calcular según el tipo de la variable.
Discreta -> entropía. Continua -> varianza y AUC (si se proporcionan clases).}
\usage{
.calcular_metricas_variable(columna, clases, verbose = 1)
}
\arguments{
\item{columna}{Variable a analizar.}

\item{clases}{Columna binaria opcional con la que evaluar el AUC de las variables continuas.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
El resultado de la métrica según el tipo de variable (entropía, varianza o varianza+auc).
}
\description{
Decide qué métrica calcular según el tipo de la variable.
Discreta -> entropía. Continua -> varianza y AUC (si se proporcionan clases).
}
