% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_de_variables.r
\name{.calcular_varianza}
\alias{.calcular_varianza}
\title{Calcula la varianza de una variable continua.
Mide cuánto se dispersan los valores respecto a su media. Formula: Var(X) = (1/n) * sumatorio((xi - media)^2).}
\usage{
.calcular_varianza(atributo_clm, verbose)
}
\arguments{
\item{atributo_clm}{Vector de valores continuos de la variable objetivo.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
El valor de la varianza calculada.
}
\description{
Calcula la varianza de una variable continua.
Mide cuánto se dispersan los valores respecto a su media. Formula: Var(X) = (1/n) * sumatorio((xi - media)^2).
}
