% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizacion_estandarizacion.r
\name{.estandarizar_variable}
\alias{.estandarizar_variable}
\title{Estandariza una variable numérica con puntuación z (media 0, desviación 1).}
\usage{
.estandarizar_variable(columna, verbose = 1)
}
\arguments{
\item{columna}{Variable continua a estandarizar.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
La variable transformada con media 0 y desviación estándar 1.
}
\description{
Estandariza una variable numérica con puntuación z (media 0, desviación 1).
}
