% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizacion_estandarizacion.r
\name{.normalizar_variable}
\alias{.normalizar_variable}
\title{Normaliza una variable numérica a escala \link{0, 1} (min-max).}
\usage{
.normalizar_variable(columna, verbose = 1)
}
\arguments{
\item{columna}{Variable continua a normalizar.}

\item{verbose}{Nivel de verbosidad.}
}
\value{
La variable transformada en el rango \link{0, 1}.
}
\description{
Normaliza una variable numérica a escala \link{0, 1} (min-max).
}
