% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{entropia}
\alias{entropia}
\title{Funcion para calcular la entropia de un vector o dataframe}
\usage{
entropia(x)
}
\arguments{
\item{df:}{vector o dataframe de datos}
}
\value{
: valor de la entropia o vector con la entropia de cada columna
}
\description{
Funcion para calcular la entropia de un vector o dataframe
}
