% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrado.R
\name{filtrar}
\alias{filtrar}
\title{Funcion para filtrar un dataframe por entropia o varianza. Se eliminan todas aquellas
columnas que no cumplan la condicion indicada. Si se quiere filtrar por entropia, se debe 
indicar el parametro Entropy como True, si se quiere filtrar por varianza, se debe indicar
como False.}
\usage{
filtrar(df, Entropy = TRUE, condition = "", value = 0)
}
\arguments{
\item{df:}{dataframe}

\item{Entropy:}{booleano para filtrar por entropia o varianza (por defecto True)}

\item{condition:}{string con la condicion para filtrar (<, >, <=, >=, ==)}

\item{valueV:}{valor para filtrar}
}
\value{
: dataframe filtrado
}
\description{
Funcion para filtrar un dataframe por entropia o varianza. Se eliminan todas aquellas
columnas que no cumplan la condicion indicada. Si se quiere filtrar por entropia, se debe 
indicar el parametro Entropy como True, si se quiere filtrar por varianza, se debe indicar
como False.
}
