% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{informacion_mutua}
\alias{informacion_mutua}
\title{Funcion que calcula la matriz de informacion mutua de un dataframe entre las variables categóricas.}
\usage{
informacion_mutua(df)
}
\arguments{
\item{df:}{dataframe}
}
\value{
: lista con el df de informacion mutua entre las variables categoricas y la matriz de informacion mutua
}
\description{
Funcion que calcula la matriz de informacion mutua de un dataframe entre las variables categóricas.
}
