% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{metricas}
\alias{metricas}
\title{Funcion que realiza el cálculo de métricas para los atributos de un dataset: varianza y AUC para las variables contínuas y entropía para las discretas.
Esta funcion reconoce automaticmente el tipo de variable de cada columna del dataframe y calcula la métrica correspondiente.}
\usage{
metricas(df, bin_class = NULL, Roc_curve = FALSE)
}
\arguments{
\item{df:}{dataframe de datos}

\item{bin_class:}{nombre de la columna con la clasificacion binaria}

\item{Roc_curve:}{booleano que indica si se desea graficar la curva ROC}
}
\value{
: lista de 2 dataframes, el primero con las metricas de las variables contínuas y el segundo con las metricas de las variables discretas
}
\description{
Funcion que realiza el cálculo de métricas para los atributos de un dataset: varianza y AUC para las variables contínuas y entropía para las discretas.
Esta funcion reconoce automaticmente el tipo de variable de cada columna del dataframe y calcula la métrica correspondiente.
}
