% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plot_mat_heatmap}
\alias{plot_mat_heatmap}
\title{Funcion para graficar la matriz de correlaciones e informacion mutua}
\usage{
plot_mat_heatmap(df)
}
\arguments{
\item{df:}{df de correlaciones e informacion mutua}
}
\description{
Funcion para graficar la matriz de correlaciones e informacion mutua
}
