% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Función para representar la curva ROC.}
\usage{
plot_roc_curve(tprl, fprl, zoom = FALSE)
}
\arguments{
\item{tprl:}{lista de valores de verdaderos positivos}

\item{fprl:}{lista de valores de falsos positivos}

\item{zoom:}{booleano para indicar si se quiere hacer zoom en la gráfica}
}
\description{
Función para representar la curva ROC.
}
