% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{t_discretizarEW}
\alias{t_discretizarEW}
\title{Funcion para discretizar un dataframe con el metodo de equal width}
\usage{
t_discretizarEW(x, num_bins, labels = FALSE)
}
\arguments{
\item{x:}{dataframe}

\item{num_bins:}{número de intervalos que queremos obtener}

\item{labels:}{Booleano para indicar si se quiere que los intervalos se muestren en el array (True) o solo el indice del intervalo (False)}
}
\value{
: dataframe discretizado
}
\description{
Funcion para discretizar un dataframe con el metodo de equal width
}
