% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{t_entropia}
\alias{t_entropia}
\title{Funcion para calcular la entropia de un dataframe}
\usage{
t_entropia(x)
}
\arguments{
\item{df:}{dataframe de datos}
}
\value{
: vector con la entropia de cada columna
}
\description{
Funcion para calcular la entropia de un dataframe
}
