% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-est.R
\name{t_estandarizar}
\alias{t_estandarizar}
\title{Funcion para estandarizar un dataframe}
\usage{
t_estandarizar(x)
}
\arguments{
\item{df:}{dataframe de datos}
}
\value{
: dataframe estandarizado
}
\description{
Funcion para estandarizar un dataframe
}
