% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-est.R
\name{t_normalizar}
\alias{t_normalizar}
\title{Funcion para normalizar un dataframe}
\usage{
t_normalizar(x)
}
\arguments{
\item{df:}{dataframe de datos}
}
\value{
: dataframe normalizado
}
\description{
Funcion para normalizar un dataframe
}
