% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{t_varianza}
\alias{t_varianza}
\title{Funcion para calcular la varianza de un dataframe}
\usage{
t_varianza(x)
}
\arguments{
\item{df:}{dataframe de datos}
}
\value{
: vector con la varianza de cada columna
}
\description{
Funcion para calcular la varianza de un dataframe
}
