% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{v_cor}
\alias{v_cor}
\title{Funcion que calcula la correlacion de dos vectores.}
\usage{
v_cor(x, y)
}
\arguments{
\item{x:}{vector}

\item{y:}{vector}
}
\value{
: correlacion de los vectores
}
\description{
Funcion que calcula la correlacion de dos vectores.
}
