% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{v_discretizarEW}
\alias{v_discretizarEW}
\title{Funcion para discretizar un vector con el metodo de equal width}
\usage{
v_discretizarEW(x, num_bins, labels = FALSE)
}
\arguments{
\item{x:}{vector}

\item{num_bins:}{número de intervalos que queremos obtener}

\item{labels:}{Booleano para indicar si se quiere que los intervalos se muestren en el array (True) o solo el indice del intervalo (False)}
}
\value{
: vector discretizado
}
\description{
Funcion para discretizar un vector con el metodo de equal width
}
