% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{v_entropia}
\alias{v_entropia}
\title{Funcion para calcular la entropia de un vector}
\usage{
v_entropia(x)
}
\arguments{
\item{x:}{vector de datos}
}
\value{
: entropia del vector
}
\description{
Funcion para calcular la entropia de un vector
}
