% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-est.R
\name{v_estandarizar}
\alias{v_estandarizar}
\title{Funcion para estandarizar un vector}
\usage{
v_estandarizar(x)
}
\arguments{
\item{x:}{vector de datos}
}
\value{
: vector estandarizado
}
\description{
Funcion para estandarizar un vector
}
