% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{v_mutua}
\alias{v_mutua}
\title{Funcion que calcula la informacion mutua de dos vectores.}
\usage{
v_mutua(x, y)
}
\arguments{
\item{x:}{vector}

\item{y:}{vector}
}
\value{
: informacion mutua de los vectores
}
\description{
Funcion que calcula la informacion mutua de dos vectores.
}
