% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-est.R
\name{v_normalizar}
\alias{v_normalizar}
\title{Funcion para normalizar un vector}
\usage{
v_normalizar(x)
}
\arguments{
\item{x:}{vector de datos}
}
\value{
: vector normalizado
}
\description{
Funcion para normalizar un vector
}
