% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{v_varianza}
\alias{v_varianza}
\title{Función para calcular la varianza de un vector}
\usage{
v_varianza(x)
}
\arguments{
\item{x:}{vector de datos}
}
\value{
: varianza del vector
}
\description{
Función para calcular la varianza de un vector
}
