% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas.R
\name{varianza}
\alias{varianza}
\title{Funcion para calcular la varianza de un vector o dataframe}
\usage{
varianza(df)
}
\arguments{
\item{df:}{vector o dataframe de datos}
}
\value{
: valor de la varianza o vector con la varianza de cada columna
}
\description{
Funcion para calcular la varianza de un vector o dataframe
}
