% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizar}
\alias{discretizar}
\title{Funcion para disctetizar un vector de datos}
\usage{
discretizar(x, bins, labels = FALSE)
}
\arguments{
\item{x:}{vector}

\item{bins:}{puntos de corte}

\item{labels:}{Booleano para indicar si se quiere que los intervalos se muestren en el array (True) o solo el indice del intervalo (False)}
}
\value{
: vector de datos discretizado
}
\description{
Funcion para disctetizar un vector de datos
}
