% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizar_anchura}
\alias{discretizar_anchura}
\title{Discretiza una variable numérica en intervalos de igual anchura.}
\usage{
discretizar_anchura(columna_variable, n_intervalos, verbose = 1)
}
\arguments{
\item{columna_variable}{Variable numérica a discretizar.}

\item{n_intervalos}{Número de intervalos deseados. Si es NULL, se calcula como el 50\% del número de casos (mínimo 1).}

\item{verbose}{Nivel de verbosidad.}
}
\value{
Una lista con los pares de límites de intervalo \link{limite_inferior, limite_superior}.
}
\description{
Discretiza una variable numérica en intervalos de igual anchura.
}
