% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizar_frecuencia}
\alias{discretizar_frecuencia}
\title{Discretiza una variable numérica en intervalos de igual frecuencia.}
\usage{
discretizar_frecuencia(columna_variable, n_intervalos, frecuencia, verbose = 1)
}
\arguments{
\item{columna_variable}{Variable numérica a discretizar.}

\item{n_intervalos}{Número de intervalos deseados. Es exclusivo con "frecuencia".}

\item{frecuencia}{Número de casos por intervalo. Es exclusivo con "n_intervalos".}

\item{verbose}{Nivel de verbosidad.}
}
\value{
Una lista con los pares de límites de intervalo \link{limite_inferior, limite_superior}.
}
\description{
Discretiza una variable numérica en intervalos de igual frecuencia.
}
