% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-est.R
\name{estandarizar}
\alias{estandarizar}
\title{Funcion que estandariza un dataframe o un vector.}
\usage{
estandarizar(df)
}
\arguments{
\item{x:}{vector o dataframe}
}
\value{
: vector o dataframe estandarizado
}
\description{
Funcion que estandariza un dataframe o un vector.
}
