% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm-est.R
\name{normalizar}
\alias{normalizar}
\title{Funcion que normaliza un dataframe o un vector.}
\usage{
normalizar(x)
}
\arguments{
\item{x:}{vector o dataframe de pandas}
}
\value{
: vector o dataframe normalizado
}
\description{
Funcion que normaliza un dataframe o un vector.
}
